Separate unit tests and integration tests
