+++ coral-tflite-delegate/src/index_test.ts
@@ -15,49 +15,33 @@
  * =============================================================================
  */
 
-import {CoralDelegate} from './index';
-import {loadTFLiteModel} from 'tfjs-tflite-node';
-import type {TFLiteModel} from 'tfjs-tflite-node/dist/tflite_model';
-import * as fs from 'fs';
-import * as tf from '@tensorflow/tfjs-core';
 import '@tensorflow/tfjs-backend-cpu';
-import * as jpeg from 'jpeg-js';
+import {CoralDelegate} from './index';
 
 describe('coral delegate', () => {
-  const modelPath = './test_model/mobilenet_v2_1.0_224_inat_bird_quant_edgetpu.tflite';
-  let model: TFLiteModel;
-  let parrot: tf.Tensor;
-  let labels: string[];
-
-  beforeEach(async () => {
-    model = await loadTFLiteModel(modelPath, {
-      delegates: [new CoralDelegate([['device', 'usb']])],
-    });
-
-    // Load the input image of a parrot.
-    const parrotJpeg = jpeg.decode(
-      fs.readFileSync('./test_model/parrot-small.jpg'));
-
-    // Create an RGB array of the parrot's pixels (no Alpha channel).
-    const {width, height, data} = parrotJpeg;
-    const parrotRGB = new Uint8Array(width * height * 3);
-    for (let i = 0; i < width * height; i++) {
-      const i3 = i * 3;
-      const i4 = i * 4;
-      parrotRGB[i3] = data[i4];
-      parrotRGB[i3 + 1] = data[i4 + 1];
-      parrotRGB[i3 + 2] = data[i4 + 2];
-    }
-
-    parrot = tf.tensor(parrotRGB, [1, 224, 224, 3]);
-    labels = fs.readFileSync('./test_model/inat_bird_labels.txt', 'utf-8')
-      .split('\n');
+  it('has the name "CoralDelegate"', () => {
+    expect(new CoralDelegate().name).toEqual('CoralDelegate');
   });
 
-  it('runs a coral model (will fail without coral device)', () => {
-    const prediction = model.predict(parrot);
-    const argmax = tf.argMax(prediction as tf.Tensor, 1);
-    const label = labels[argmax.dataSync()[0]];
-    expect(label).toEqual('Ara macao (Scarlet Macaw)');
+  it('stores options', () => {
+    const options: Array<[string, string]> = [['foo', 'bar'], ['123', '456']];
+    const coralDelegate = new CoralDelegate(options);
+    expect(coralDelegate.options).toEqual(options);
+  });
+
+  it('allows manually setting lib path', () => {
+    const libPath = 'some lib path';
+    const coralDelegate = new CoralDelegate([], libPath);
+    expect(coralDelegate.node.path).toEqual(libPath);
+  });
+
+  it('sets the lib path automatically based on platorm', () => {
+    const coralLinux = new CoralDelegate([], undefined, 'linux');
+    const coralMac = new CoralDelegate([], undefined, 'darwin');
+    const coralWindows = new CoralDelegate([], undefined, 'win32');
+
+    expect(coralLinux.node.path).toEqual('libedgetpu.so.1');
+    expect(coralMac.node.path).toEqual('libedgetpu.1.dylib');
+    expect(coralWindows.node.path).toEqual('edgetpu.dll');
   });
 });
